mission-control: extract the cruise-speed control into a component
The speedometer button, its menu and the debounced speed command were
duplicated between the mission control widget and its mini-widget,
differing only in the icon size, so a single component now serves both
through an icon-class prop. It also clears the debounce timer on unmount
and logs the settled speed as a user action, neither of which the copies
did.
