+++ src/components/mission-planning/CruiseSpeedControl.vue
@@ -1,0 +1,79 @@
+<template>
+  <div class="relative flex items-center -mt-[2px]">
+    <v-menu :close-on-content-click="false" location="top" offset="8">
+      <template #activator="{ props: speedProps }">
+        <v-tooltip location="top" open-delay="800" text="Cruise speed">
+          <template #activator="{ props: speedTooltipProps }">
+            <v-btn
+              v-bind="{ ...speedProps, ...speedTooltipProps }"
+              size="x-small"
+              icon="mdi-speedometer"
+              variant="text"
+              :class="iconClass"
+              :disabled="!vehicleStore.isVehicleOnline"
+            />
+          </template>
+        </v-tooltip>
+      </template>
+      <div class="flex flex-col p-3 rounded-lg w-[210px] text-white" :style="interfaceStore.globalGlassMenuStyles">
+        <div class="flex justify-between items-center mb-1 text-xs">
+          <span>Cruise speed</span>
+          <span class="font-bold">{{ liveCruiseSpeed.toFixed(1) }} m/s</span>
+        </div>
+        <v-slider
+          v-model="liveCruiseSpeed"
+          :min="0.1"
+          :max="5"
+          :step="0.1"
+          color="white"
+          density="compact"
+          hide-details
+          @update:model-value="handleCruiseSpeedInput"
+        />
+      </div>
+    </v-menu>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { onBeforeUnmount, ref, watch } from 'vue'
+
+import { openSnackbar } from '@/composables/snackbar'
+import { useAppInterfaceStore } from '@/stores/appInterface'
+import { useMainVehicleStore } from '@/stores/mainVehicle'
+import { useMissionStore } from '@/stores/mission'
+
+withDefaults(
+  defineProps<{
+    /**
+     * Classes applied to the speedometer icon button
+     */
+    iconClass?: string
+  }>(),
+  { iconClass: 'text-[18px]' }
+)
+
+const interfaceStore = useAppInterfaceStore()
+const missionStore = useMissionStore()
+const vehicleStore = useMainVehicleStore()
+
+const liveCruiseSpeed = ref<number>(Number(missionStore.cruiseSpeed))
+watch(
+  () => missionStore.cruiseSpeed,
+  (newSpeed) => (liveCruiseSpeed.value = Number(newSpeed))
+)
+
+// Debounce live speed commands so dragging the slider doesn't flood the vehicle with DO_CHANGE_SPEED.
+let cruiseSpeedDebounce: ReturnType<typeof setTimeout> | undefined
+const handleCruiseSpeedInput = (value: number): void => {
+  if (cruiseSpeedDebounce) clearTimeout(cruiseSpeedDebounce)
+  cruiseSpeedDebounce = setTimeout(() => {
+    logUserAction(`Set the cruise speed to ${value.toFixed(1)} m/s`)
+    missionStore.applyCruiseSpeed(value).catch((err) => {
+      openSnackbar({ message: `Failed to set cruise speed: ${(err as Error).message}`, variant: 'error' })
+    })
+  }, 300)
+}
+
+onBeforeUnmount(() => clearTimeout(cruiseSpeedDebounce))
+</script>
